refactor(AuthController.java): 토큰 발급 요청에 대한 컨트롤러와 서비스 변경
