+++ src/main/java/woowacourse/auth/ui/AuthController.java
@@ -1,4 +1,28 @@
 package woowacourse.auth.ui;
 
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import woowacourse.auth.application.AuthService;
+import woowacourse.auth.dto.TokenRequest;
+import woowacourse.auth.dto.TokenResponse;
+import woowacourse.shoppingcart.application.dto.SignInDto;
+
+@Controller
+@RequestMapping("/api")
 public class AuthController {
+
+    private final AuthService authService;
+
+    public AuthController(AuthService authService) {
+        this.authService = authService;
+    }
+
+    @PostMapping("/customer/authentication/sign-in")
+    public ResponseEntity<TokenResponse> signIn(@RequestBody final TokenRequest tokenRequest) {
+        final TokenResponse tokenResponse = authService.signIn(SignInDto.fromTokenRequest(tokenRequest));
+        return ResponseEntity.ok().body(tokenResponse);
+    }
 }
